# backend/swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/reports": {
      "post": {
        "operationId": "ReportsController_createReport",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateReportDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Report generation request received."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a new report (PDF or CSV)",
        "tags": [
          "Reports"
        ]
      },
      "get": {
        "operationId": "ReportsController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Reports list retrieved successfully."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get all reports with filters",
        "tags": [
          "Reports"
        ]
      }
    },
    "/reports/pdf": {
      "post": {
        "operationId": "ReportsController_createPdfReport",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePdfReportDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "PDF report generation request received."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a PDF report asynchronously",
        "tags": [
          "Reports"
        ]
      }
    },
    "/reports/csv": {
      "post": {
        "operationId": "ReportsController_createCsvReport",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCsvReportDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "CSV report generation request received."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a CSV report asynchronously",
        "tags": [
          "Reports"
        ]
      }
    },
    "/reports/{id}": {
      "get": {
        "operationId": "ReportsController_findById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Report ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Report details retrieved."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get report details by ID",
        "tags": [
          "Reports"
        ]
      },
      "delete": {
        "operationId": "ReportsController_deleteReport",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Report ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Report deleted successfully."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete a report",
        "tags": [
          "Reports"
        ]
      }
    },
    "/reports/{id}/download": {
      "get": {
        "operationId": "ReportsController_downloadReport",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Report ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "File stream returned."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Download generated report file by ID",
        "tags": [
          "Reports"
        ]
      }
    },
    "/reports/{id}/retry": {
      "post": {
        "operationId": "ReportsController_retryReport",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Report ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Retrying report generation."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Retry generating a failed report",
        "tags": [
          "Reports"
        ]
      }
    },
    "/leaderboard": {
      "get": {
        "operationId": "LeaderboardController_getLeaderboard",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Developer leaderboard retrieved."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get developer leaderboard ranked by DQS",
        "tags": [
          "Leaderboard"
        ]
      }
    },
    "/leaderboard/teams": {
      "get": {
        "operationId": "LeaderboardController_getTeamLeaderboard",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Team leaderboard retrieved."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get team leaderboard ranked by average DQS",
        "tags": [
          "Leaderboard"
        ]
      }
    },
...
    "/teams": {
      "post": {
        "operationId": "TeamsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTeamDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Team created successfully"
          },
          "403": {
            "description": "User does not have permission to create teams"
          },
          "409": {
            "description": "Team with this name already exists"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a new team",
        "tags": [
          "Teams"
        ]
      },
      "get": {
        "operationId": "TeamsController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "List of teams"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get all teams for the current organization",
        "tags": [
          "Teams"
        ]
      }
    },
    "/teams/leaderboard": {
      "get": {
        "operationId": "TeamsController_getLeaderboard",
        "parameters": [
          {
            "name": "projectId",
            "required": false,
            "in": "query",
            "description": "Filter by project ID",
            "schema": {
              "example": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Team leaderboard",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TeamLeaderboardResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get team leaderboard ranked by DQS",
        "tags": [
          "Teams"
        ]
      }
    },
    "/teams/{id}": {
      "get": {
        "operationId": "TeamsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Team ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Team details"
          },
          "404": {
            "description": "Team not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get team by ID",
        "tags": [
          "Teams"
        ]
      },
      "patch": {
        "operationId": "TeamsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Team ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTeamDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Team updated successfully"
          },
          "403": {
            "description": "User does not have permission to update this team"
          },
          "404": {
            "description": "Team not found"
          },
          "409": {
            "description": "Team with this name already exists"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update team",
        "tags": [
          "Teams"
        ]
      },
      "delete": {
        "operationId": "TeamsController_delete",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Team ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Team deleted successfully"
          },
          "403": {
            "description": "User does not have permission to delete this team"
          },
          "404": {
            "description": "Team not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete team",
        "tags": [
          "Teams"
        ]
      }
    },
...
  },
  "info": {
    "title": "My API",
    "description": "API documentation",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "CreateReportDto": {
        "type": "object",
        "properties": {}
      },
      "CreatePdfReportDto": {
        "type": "object",
        "properties": {}
      },
      "CreateCsvReportDto": {
        "type": "object",
        "properties": {}
      },
...
      "CreateTeamDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Team name",
            "example": "Backend Team",
            "minLength": 1,
            "maxLength": 100
          },
          "description": {
            "type": "string",
            "description": "Team description",
            "example": "Responsible for backend services and APIs",
            "maxLength": 500
          }
        },
        "required": [
          "name"
        ]
      },
      "TeamLeaderboardEntryDto": {
        "type": "object",
        "properties": {
          "teamId": {
            "type": "string",
            "description": "Team ID"
          },
          "name": {
            "type": "string",
            "description": "Team name"
          },
          "memberCount": {
            "type": "number",
            "description": "Number of active team members",
            "example": 5
          },
          "aggregatedDQS": {
            "type": "object",
            "description": "Aggregated team DQS (weighted average by commit count)",
            "example": 75.5,
            "nullable": true
          },
          "dqsTrend": {
            "type": "object",
            "description": "DQS trend indicator (positive = improving, negative = declining)",
            "example": 2.5,
            "nullable": true
          },
          "totalCommits": {
            "type": "number",
            "description": "Total commits in the last 30 days",
            "example": 150
          },
          "status": {
            "type": "string",
            "description": "Status indicator",
            "example": "Active",
            "enum": [
              "Active",
              "No Activity"
            ]
          },
          "rank": {
            "type": "number",
            "description": "Rank position in leaderboard",
            "example": 1
          }
        },
        "required": [
          "teamId",
          "name",
          "memberCount",
          "aggregatedDQS",
          "dqsTrend",
          "totalCommits",
          "status",
          "rank"
        ]
      },
      "TeamLeaderboardResponseDto": {
        "type": "object",
        "properties": {
          "teams": {
            "description": "List of teams ranked by DQS",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TeamLeaderboardEntryDto"
            }
          },
          "totalTeams": {
            "type": "number",
            "description": "Total number of teams",
            "example": 10
          },
          "projectFilter": {
            "type": "string",
            "description": "Project filter applied",
            "example": "uuid"
          }
        },
        "required": [
          "teams",
          "totalTeams"
        ]
      },
...
      "UpdateTeamDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Team name",
            "example": "Backend Team",
            "minLength": 1,
            "maxLength": 100
          },
          "description": {
            "type": "string",
            "description": "Team description",
            "example": "Responsible for backend services and APIs",
            "maxLength": 500
          }
        }
      },
...
    }
  }
}